(MODULES-6845) Copy V2 Provider into Win32
Prior to this commit the Win32 provider in this module was a near
mirror of the existing provider in core Puppet.  In order to more
easily maintain the new module, to debug it, and to onboard team
members to maintaining it, this commit copies over the code from
the new provider into the old one. As the new provider was made
to keep in parity with the old one and the tests already existed
to verify this, the change is determined to be low risk.

In a future commit the V2 provider will be updated to manage
scheduled tasks of all compatibility modes and will be the code
which gets all new features and improvements. The legacy provider
will be frozen at a point in the near future for all changes
except for security and critical bug fixes.

This commit also updates the test suite to verify the changes and
removes unneccessary logic from the tests which were needed when
the providers did not share helper code.

However, in this iteration, the tests only run over the legacy
provider and not the V2 provider, though they are using the same
helper at this time.  In a future commit, when the V2 provider is
updated to use the new V2 helper, the test suite will again run
for both.