(MODULES-6845) TaskScheduler2Task in API2 Provider
Prior to this commit both the win32 (legacy) and api2 (new)
providers leveraged the adapter code in `TaskScheduler2V1Task`
to manage scheduled tasks.  This commit updates the API2
provider to use the TaskScheduler2Task helper, allowing us to
freeze work on the adapter code and begin to add new features
to the API2 helper.

At this point in time, only the helper has been swapped out,
the provider itself has not been meaningfully altered.